iwlwifi: update trans->hw_rev 8000 hw family format
The format of the CSR_HW_REV register has changed in 8000
HW family. To keep backwards compatibility, we store the
value of this register as usual in trans->hw_rev, only we
store it in the old format in this variable.

Change-Id: Iba29f9ad5073ac8539933731c5b00635f958905c
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/38166
Reviewed-by: Eran Harary <eran.harary@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: f3152a940960cbd7bcfc503165e47f087b9167c3