mac80211: add functions to stop and wake all queues assigned to a vif
In some cases we may want to stop the queues of a single vif (for
instance during a channel-switch).  Add a function that stops all the
queues that are assigned to a vif.  If a queue is assigned to more
than one vif, the corresponding netdev subqueue of the other vif(s)
will also be stopped.  If the HW doesn't set the
IEEE80211_HW_QUEUE_CONTROL flag, then all queues are stopped.

Also add a corresponding function to wake the queues of a vif back.

Change-Id: I9c7ad9902e9ff36f70ca5028c52c733d2b4de8ae
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/38418
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 210afa7cbbff42f2b05ccc20c78eae1993ccbaa4