[MERGEUPSTREAM] mac80211: introduce refcount for queue_stop_reasons
Sometimes different vifs may be stopping the queues for the same
reason (e.g. when several interfaces are performing a channel switch).
Instead of using a bitmask for the reasons, use an integer that holds
a refcount instead.  In order to keep it backwards compatible,
introduce a boolean in some functions that tell us whether the queue
stopping should be refcounted or not.  For now, use not refcounted for
all calls to keep it functionally the same as before.

Conflicts:
	net/mac80211/cfg.c
	net/mac80211/ieee80211_i.h
	net/mac80211/iface.c
	net/mac80211/mlme.c
	net/mac80211/tdls.c
	net/mac80211/debugfs.c

Change-Id: Ifea1506656fc1480339fe26df1ec4bcf35605900
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/38826
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: a7942c592336c3d8d0d4a85f290712bfbce62b3c