iwlwifi: testmode: fix dbgc trace retrieval in 8000 hw b-step
The behavior of the wrptr of the DBGC has changed in the
B-step, so this updates accordingly. It also adds an
additional check in the code to make sure that even if the
wrptr is read wrong - no out-of-bounds occurs.

Change-Id: Ibd168baad0de9754bc182218af854454116df751
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/38468
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 7830b13f11837bdab0687f7a88e877211d70a8a7