mac80211: add extended channel switching capability if the driver supports CSA
The Extended Channel Switching capability bit in the extended
capabilities element must be set if the driver supports CSA on
non-beaconing interfaces.

Since this capability needs to be set during driver registration, the
extended_capabiliities global variable needs to be moved to the local
structure so that it can be modified.

Change-Id: I657ff40e839833c52e653ac226e4cf0e09c7ea53
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/38582
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins

x-iwlwifi-stack-dev: 023488a95652746155b82eaf84e3687eee463e34