iwlwifi: mvm: read the mac address in family 8000
in family 8000 products the mac address in the otp could be in both:
- WFPM address
- PCIE address
in sdio product driver should read it from the WFPM, in pcie product
it driver should read it from the PCIe location.
but: this change is relevant only from otp version 0xE08 and above.
also: in version 0xE08 the bytes order are wrong and drive should
switch the bytes.

Change-Id: I5469df4d4cb083e294500e41b06be38a38aef39a
Signed-off-by: Eran Harary <eran.harary@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/38459
Tested-by: IWL Jenkins
Reviewed-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 8f75f0ac7d19040de2021aa6aab54613dd14fc82