mac80211: fix dtim period propagation to the driver
In case the beacon is heard before we try to associate,
the dtim_period was not communicated to the low level
driver. This led to disconnections with iwlmvm which needs
to know the dtim_period before it can move its firmware to
'associated' state. The following print would appear in the
log prior the disconnection.

iwlwifi: No association and the time event is over already...

Note that we usually hear the beacon only after the
association because we associate based on the probe
responses from the scan. This is why this bug has not been
discovered until now.

https://bugzilla.kernel.org/show_bug.cgi?id=78101

Change-Id: I1b39d2755f8f892ba9da904c6903b122ceb00d71
Cc: <stable@vger.kernel.org> [3.9+]
Fixes: c65dd1477b6f ("mac80211: inform the driver about update of dtim_period")
Reported-by: Johannes Stezenbach <js@sig21.net>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/39481
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 505c50ece615e03319dbee5df0f4ffb27f6c1209