mac80211: track TDLS initiator internally
Infer the TDLS initiator and track it in mac80211 via a STA flag. This
avoids breaking old userspace that doesn't pass it via nl80211 APIs.

The only case where userspace will need to pass the initiator is when the
STA is removed due to unreachability before a teardown packet is sent.
Support for unreachability was only recently added to wpa_supplicant, so
it won't be a problem in practice.

Change-Id: Ib0f3c5fbf5500b6a4c6e444718340541b267b207
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/39911
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: d09d6eb2c2a8db2ae19c7713badc2f5fdc166551