iwlwifi: mvm: Define AUX ROC Command
Add new AUX ROC command that is orioinally made for HS2.0 purposes.
It is used when sending ANQP requests on a specific channel is needed.

This command requests the ucode to trigger a time event and remain on a certain
channel for a given duration.
Triggering the time event is done by using the Aux Framework in the FW, and
makes use of the Aux station (similarly to scan).

The ucode responds with a response command, that indicates the unique id for the
time event that will be triggered.

Change-Id: I0f7a6affd60bc0e41dbf12c4dea7574ae86b7f22
Signed-off-by: Ariej Marjieh <ariej.marjieh@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/37389
Tested-by: IWL Jenkins
Reviewed-by: Dor Shaish <dor.shaish@intel.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 3b695427c3ed57efdeda7bd5c9a341bfb24d4948