iwlwifi: mvm: Add Aux ROC request/response flow
The driver receives an action frame from the wpa_supplicant via nl80211 that
requests to be sent offchannel

The driver sends to the ucode the Aux ROC command (0x53), to request to remain
on the channel via the Aux framework with the Aux station.

The ucode responds with a similar command, and supplies the unique id.

Once the time event begins, the driver puts in the Aux queue the action frame
that was given by the wpa_supplicant, and the ucode sends it.

Change-Id: I7c4021a0fbcc73101667487858d13a6eb1ac5918
Signed-off-by: Ariej Marjieh <ariej.marjieh@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/37390
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: b26dc5ca9af64687be259636be55d10f8713ca2d