mac80211: don't duplicate station QoS capability data
We currently track the QoS capability twice: for all peer stations
in the WLAN_STA_WME flag, and for any clients associated to an AP
interface separately for drivers in the sta->sta.wme field.

Remove the WLAN_STA_WME flag and track the capability only in the
driver-visible field, getting rid of the limitation that the field
is only valid in AP mode.

Change-Id: I492f3f437e0ecc0f64d76e3e5d66d41e0f24dd80
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/41163
Tested-by: IWL Jenkins
Reviewed-by: ArikX Nemtsov <arik@wizery.com>

x-iwlwifi-stack-dev: e1b3a13dde4436aebd79692a7c4966b4a2774f8a