iwlwifi: mvm: set MAC_FILTER_IN_BEACON correctly for STA/P2P client
In commit cad3f08c (iwlwifi: mvm: enable MAC_FILTER_IN_BEACON when
forced_assoc_off is set) the code to set the MAC_FILTER_IN_BEACON flag
was accidentally moved to the main block of the if statement, while it
should be in the else block instead.  Move it to the right place.

Fixes: cad3f08c23de ("iwlwifi: mvm: enable MAC_FILTER_IN_BEACON when forced_assoc_off is set")
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---
Note that the commit ID mentioned in the commit message refers to the
commit ID from upstream.  The local commit ID is c65369c3.

Change-Id: I456662a51a84196fa542781bd8ba628a40955119
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/41648
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 966e90ba82b21446b66f47cf4352f26ae4d6b80d