mac80211: fix broken use of VHT/20Mhz with some APs [SQUASH]
commit 7a5b3cf2 "mac80211: disable 40MHz support in case of 20MHz AP"
broke working VHT in 20Mhz with APs like Netgear R6300v2 which
do not publish support for 40Mhz but allow use of VHT in 20Mhz.
The break is because VHT is disabled once no HT cap doesn't indicate
support for 40Mhz. This causes the assoc request to be sent without
any VHT IE and the association is only HT due to this.

For more details check out commit 4a817aa7
"mac80211: allow VHT with peers not capable of 40MHz"

Change-Id: I6794c7abd18aad4fb1d514047ce05e22638e567f
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
This should be squashed with Eliad's fix if it hasn't been upstreamed yet
Reviewed-on: https://gerrit.rds.intel.com/42237
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 3f2d48fb607a71bafa2c2b71f0bff90589554aa3