iwlwifi: mvm: correct firmware disassoc command sequence
The firmware would like to have a MAC context (unassoc)
before the AP station is removed (we do this) but would
like to keep the BSSID until after it is removed, so we
need to send two commands - one with the BSSID before
and one without the BSSID after.

In order to do this, we need to store the BSSID as it
will have been cleared by mac80211 by the time we get
notified of the disassociation. Also pass it around as
an override to the various functions needing it, and
keep taking it from the mac80211 data otherwise. This
avoids having to keep track of the BSSID in all modes.

Change-Id: Ia49ae36e1b69fed5b1733ef83a7f80259da7393c
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/42316
Tested-by: IWL Jenkins
Tested-by: Assaf Azulay <assaf.azulay@intel.com>
Reviewed-by: Dor Shaish <dor.shaish@intel.com>
Reviewed-by: Assaf Azulay <assaf.azulay@intel.com>

x-iwlwifi-stack-dev: f2333f25cd6f5cb2f088761b37b7185552bc11d3