iwlwifi: mvm: choose an initial tx rate based on rssi conditions
Improve the initial tx rate and antenna selection to be based on
the rssi of the last rx. This avoids starting at the lowest legacy
rate always and requiring more tx traffic to "climb" up the rates.

Change-Id: I4ac38ce0025fd66c4acf67bda7b74e73a168b435
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
The table and behavior should conform to TLC SAS 4.2.6
Except I'm not basing this only on beacons.
This should help with passing a problematic WFA certification automatic
test which doesn't expect initial traffic to be in legacy rates and
is looking for VHT traffic.
Reviewed-on: https://gerrit.rds.intel.com/42388
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 4d2462f60d6b4eb67a48fb7f57c549d213936b8e