iwlwifi: dnt: allow dma trace retrieval also on hw error
On HW error, the monitor's write pointer read will have an
error value and thus currently fails the trace retrieval.
This is problematic, since we want to be able to pull the
traces anyway.

Since the write pointer is just needed for buffer reordering,
if we can't read it correctly - just return the buffer's
content as-is without reordering.

Change-Id: I8d9c70afe29e6ffcb6bc4d2d3d44c8bf3f25e45f
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/42476
Reviewed-by: IdanX Kahlon <idanx.kahlon@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>

x-iwlwifi-stack-dev: 0848ebd580da065f6f2e78cb4a20c2b624bba415