iwlwifi: mvm: Fix skip over dtim configuration in d0i3 [BUGFIX]
There was some confusion concerning the units of the beacon interval.
The driver assumed that it was in msec where it was in TU - so fix that.
Skip over dtim was capped by 300TU where it should be by 306TU.
It should also be subjected to several conditions:
Not a DFS channel, dtim period < 10, and the multicast wake-lock
is off.  Concerning multicast lock - there is an implementation gap
in the supplicant, so just leave a TODO.

---
MWG100215093

Change-Id: Ic17b486415631c40a027aab6719234d57abd006f
Signed-off-by: Avri Altman <avri.altman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/43068
Tested-by: IWL Jenkins
Reviewed-by: EliadX Peller <eliad@wizery.com>
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 0760d9a79694e8430591939e51127f4777d57752