iwlwifi: mvm: rs: remove max_rate_idx
max_rate_idx constraint is deprecated and it's handling is
faulty as well as it is relevant only for legacy rates but
was considered in HT/VHT. In most cases there was no side effect
as max_rate_idx was set to -1 but in certain cases like P2P
it got set to an actual rate idx which would limit the maximum
rate in HT/VHT by mistake.
max_rate_idx should be replaced by the masks fields but for
now remove it.

Change-Id: I32f1388e97f113dccebfae845ec6c1795e96c928
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
A patch implementing rate_idx_mask and rate_idx_mcs_mask
would be submitted but this was causing us trouble...
Reviewed-on: https://gerrit.rds.intel.com/43216
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 920ccbf7cb8d996d35dc50c389d47137c6f21426