[NOUPSTREAM] iwlwifi: mvm/tcm: fix dynamic quota calculations
The advanced quota calculation erroneously used the "load"
value (which is just an enum indicating low/medium/high)
rather than the percentage of airtime.

To fix this, keep track of the elapsed time and calculate
the percentage in the quota code. Also take the TCM lock
to get a consistent view of the data.

Change-Id: Id872a6bc2a6f93462567b4ec6c72ebe782388799
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/43903
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: b1afe219329fc3d7f15fe3edede4148e645a6684