iwlwifi: mvm: rs: refactor to allow direct rs updating
Enable RS to get updated directly via iwl_mvm_rs_tx_status
which will be called by the driver and not via mac80211
rate control. This is required for a following patch to
report on BAs which do not cause the BA window to progress.
Also this is a first step in decoupling rs from the
mac80211 rate control.

Change-Id: Id047a605ed5f840f78eea91274b45967dd9147a4
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/43973
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 551e5e819e8185957e92857932e0eb493e659e04