mac80211: return the vif's chandef in ieee80211_cfg_get_channel()
The chandef of the channel context a vif is using may be different
than the chandef of the vif itself.  For instance, the bandwidth used
by the vif may be narrower than the one configured in the channel
context.  To avoid confusion, return the vif's chandef in
ieee80211_cfg_get_channel() instead of the chandef of the channel
context.

Change-Id: I91c08519afb3d23b0a477da8a96219b844a97c02
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---
MWG100214155
Reviewed-on: https://gerrit.rds.intel.com/44931
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

x-iwlwifi-stack-dev: 2faf1f9b9b4a4d5f2937887e7b97d6b36aee785d